backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
yfinance==0.2.38
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
tensorflow==2.16.1
ta==0.11.0
joblib==1.4.2
python-dotenv==1.0.1
gunicorn==22.0.0
